requirements.txt
fastapi
uvicorn
numpy
opencv-python
pillow
torch
torchvision
